tiervc/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
openai==1.12.0
anthropic==0.18.1
python-dotenv==1.0.0
aiofiles==23.2.1
